+++ RELEASE_NOTES.md
@@ -3,7 +3,7 @@
 This release addresses:
 
 * Include 2024 HadUK-Grid data
-
+* Generated 'nc' files are now netCDF-4 rather than netCDF-4 classic
 
 # Release 2.14.0